agents/cluster/fluidity/cloud/requirements.txt
prometheus_client
redis
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
mlstelemetry==0.2.4
kubernetes
psutil>=5.4.5
geographiclib
flask
flask-restx
ruamel.yaml
yamllint
jsonschema
geojson
shapely
pyproj
utm
flask_cors
